Refactor(seokho/emotion): apis/emotions => apis/emotion 으로 변경 #428
